jobs_updater/requirements.txt
telethon
python-dotenv
requests
google-generativeai==0.4.1
